Handling more packet in libpcap.
When event is detected on the pcap file descriptor, we must check if
there is more than one packet, otherwise we will read only one packet
and forget all others.
To workarround this, the timeout must be less than the one given to
pcap, and we must loop over pcap_getnex() to handle all available
packets.

Signed-off-by: Baligh GUESMI <gasmibal@gmail.com>