Support NFQUEUE.
Add support for nf_queue in the link layer.
When using EVSPOT_LINK_TYPE_NFQ as type for device, this new layer will
be used, and a rule must be added to the queue created.
The defautl queue created is '0', the iptables rule is somethins like
this :
iptables -A INPUT -i eth0 -j NFQUEUE --queue-num 0

Signed-off-by: Baligh GUESMI <gasmibal@gmail.com>